feat: implement sending and displaying message on chat menu
